requirements.txt
requests
python-slugify
python-dotenv
yt-dlp
beautifulsoup4
tqdm